Refs #20352 - Backup execution plans on their deletion
When execution plan is deleted it is exported along with actions and
steps. The backup is controlled by two configuration options -
backup_deleted_plans and backup_dir. The data are exported in CSV format
one file per plans, actions and steps. The files are created under the
backup_dir and daily subdirs for easier rotations.