Dispatch delayed plan planning through connector
Before this patch, planning of delayed plans was done by the delayed executor
actor, which runs in the orchestrator process. This had several drawbacks. The
polling interval used by the delayed executor could increase when it was
performing planning, it could have required additional ActiveRecord connection
and it required the entire application environment to be loaded.

With this change, the actual planning is dispatched through connector to the
workers. To prevent planning a plan twice, a flag is set on the delayed plan
record to indicate it is already being handled. Deletion of delayed plan record
is moved to the worker performing the actual planning.

This also allows us to "plan elsewhere", kick off a task but let another process
perform the plan phase without blocking the client with planning.